refactor: move nearest-vertex/edge helpers from shape to _shape
